test: add test_transfer_task_dest_offset
